Rename project to aa-edit-data
