+++ src/aa_edit_data/edit_data.py
@@ -4,14 +4,14 @@ from pathlib import Path
 
 import typer
 
-from aa_remove_data._version import __version__
-from aa_remove_data.algorithms import (
+from aa_edit_data._version import __version__
+from aa_edit_data.algorithms import (
     apply_min_period,
     reduce_by_factor,
     remove_after_ts,
     remove_before_ts,
 )
-from aa_remove_data.archiver_data import ArchiverData
+from aa_edit_data.archiver_data import ArchiverData
 
 
 def validate_positive(value: float):
