+++ src/aa_edit_data/generate_protobuf_file.sh
