+++ tests/test_pb_tools_cli.py
@@ -6,8 +6,8 @@ from pathlib import Path
 
 from typer.testing import CliRunner
 
-from aa_remove_data import __version__
-from aa_remove_data.pb_tools import app
+from aa_edit_data import __version__
+from aa_edit_data.pb_tools import app
 
 TEST_DATA = Path("tests/test_data")
 CLI_OUTPUT = Path("tests/test_data/cli_expected_output")
@@ -23,7 +23,7 @@ def try_to_remove(filepath: PathLike):
 
 
 def test_cli_version():
-    cmd = [sys.executable, "-m", "aa_remove_data", "--version"]
+    cmd = [sys.executable, "-m", "aa_edit_data", "--version"]
     assert subprocess.check_output(cmd).decode().strip() == __version__
